Add utilities for direct * -> ast.Value transforms
Turns out we save a lot of time and resources by avoiding transforms
having to go through JSON and/or map round trips. This PR adds a bunch
of utilities to help with that, as well as some modifications to existing
code to make use of this.

More details, and benchmarks, will be provided in Regal with this merged.

Signed-off-by: Anders Eknert <anders@eknert.com>